roachtest: use a fixed seed for rng in import-cancellation roachtest
Currently, due to the randomness present in the test, there may be
significant variance in test runs. In some situations, the nature of the
random import cancellations may result in the test timing out. This
makes the test flaky.

Use a fixed seed for the test's PRNG to make the test more
deterministic.

Fix #90510.

Release note: None.